requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.2.0
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
packaging==26.0
regex==2026.2.28
requests==2.32.5
semantic-version==2.10.0
setuptools==82.0.1
setuptools-rust==1.12.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
yt-dlp==2026.3.13
